% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erinr.R
\docType{package}
\name{erinr-package}
\alias{erinr-package}
\title{erinr- functions for shuffling cards, rolling dice, taking the square root of numbers, etc.}
\description{
Miscelaneous functions for testing code in class.
}
\details{
Explain all kinds of stuff here.
}
\section{My Custom Section}{
 Add additional information here.
}
\keyword{package,dice,shuffle,roll}

