% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roll.R
\name{RP}
\alias{RP}
\title{Power Function}
\usage{
RP(x, power = 1)
}
\arguments{
\item{x}{must be numeric}

\item{power}{must be numeric}
}
\value{
Number Raised to a Power
}
\description{
Power Function
}
\examples{
RP(x=2,power=2)
}

