% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle Function for Class}
\usage{
shuffle(cards)
}
\arguments{
\item{cards}{}
}
\value{
a shuffled deck of cards to be dealt out
}
\description{
Shuffle Function for Class
}
\examples{
site <- "https://gist.githubusercontent.com/garrettgman/9629323/raw/ee5dfc039fd581cb467cc69c226ea2524913c3d8/deck.csv"
deck2 <- readr::read_csv(site)
shuffle(deck2)
}

